+++ mvc/README.md
@@ -28,7 +28,7 @@
     - JSON 자바 객체 변환 시, Jackson 라이브러리를 사용한다.
     - JSON으로 응답 시, `ContentType` -> `MediaType.APPLICATION_JSON_UTF8_VALUE`으로 변환한다.
     - model 데이터 개수 1개면 값 그대로 반환, 2개 이상이면 Map 형태 그대로 JSON으로 변환해서 반환한다.
-- [ ] Legacy MVC 제거
+- [x] Legacy MVC 제거
     - [x] app 모듈에 있는 모든 컨트롤러를 어노테이션 기반 MVC로 변경한다.
     - [x] asis 패키지에 있는 레거시 코드를 삭제해도 서비스가 정상 동작한다.
-    - [ ] DispatcherServlet를 mvc 패키지로 옮긴다.
+    - [x] DispatcherServlet를 mvc 패키지로 옮긴다.
