+++ mvc/src/main/java/com/interface21/webmvc/servlet/DispatcherServlet.java
@@ -1,11 +1,7 @@
-package com.techcourse;
+package com.interface21.webmvc.servlet;
 
-import com.interface21.webmvc.servlet.ModelAndView;
-import com.interface21.webmvc.servlet.View;
 import com.interface21.webmvc.servlet.mvc.tobe.adapter.HandlerAdapter;
 import com.interface21.webmvc.servlet.mvc.tobe.adapter.HandlerAdapterRegistry;
-import com.interface21.webmvc.servlet.mvc.tobe.adapter.HandlerExecutionAdapter;
-import com.interface21.webmvc.servlet.mvc.tobe.mapper.AnnotationHandlerMapping;
 import com.interface21.webmvc.servlet.mvc.tobe.mapper.HandlerMapping;
 import com.interface21.webmvc.servlet.mvc.tobe.mapper.HandlerMappingRegistry;
 import jakarta.servlet.ServletException;
@@ -30,8 +26,6 @@ public class DispatcherServlet extends HttpServlet {
 
     @Override
     public void init() {
-        addHandlerMapping(new AnnotationHandlerMapping(this.getClass().getPackage().getName()));
-        addHandlerAdapter(new HandlerExecutionAdapter());
     }
 
     public void addHandlerMapping(HandlerMapping handlerMapping) {
